+++ homework_15/src/main/java/ru/otus/spring/config/IntegrationConfig.java
@@ -1,0 +1,53 @@
+package ru.otus.spring.config;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.integration.annotation.IntegrationComponentScan;
+import org.springframework.integration.channel.PublishSubscribeChannel;
+import org.springframework.integration.channel.QueueChannel;
+import org.springframework.integration.config.EnableIntegration;
+import org.springframework.integration.dsl.IntegrationFlow;
+import org.springframework.integration.dsl.IntegrationFlows;
+import org.springframework.integration.dsl.MessageChannels;
+import org.springframework.integration.dsl.Pollers;
+import org.springframework.integration.scheduling.PollerMetadata;
+import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.annotation.EnableScheduling;
+import ru.otus.spring.service.AtelierService;
+
+@Configuration
+@EnableIntegration
+@IntegrationComponentScan
+@EnableAsync
+@EnableScheduling
+@RequiredArgsConstructor
+public class IntegrationConfig {
+
+    private final AtelierService atelierService;
+
+    @Bean
+    public QueueChannel clothChannel() {
+        return MessageChannels.queue(10).get();
+    }
+
+    @Bean
+    public PublishSubscribeChannel clothesChannel() {
+        return MessageChannels.publishSubscribe().get();
+    }
+
+    @Bean(name = PollerMetadata.DEFAULT_POLLER)
+    public PollerMetadata poller() {
+        return Pollers.fixedRate(100).maxMessagesPerPoll(2).get();
+    }
+
+    @Bean
+    public IntegrationFlow atelierFlow() {
+        return IntegrationFlows.from(clothChannel())
+                .split()
+                .handle(atelierService, "sew")
+                .aggregate()
+                .channel(clothesChannel())
+                .get();
+    }
+}
